News, Out of Stock, Categories, Cart AJAX
First, we introduce some other categories for the genre of the book in the bookstore.sql (database).
We added the "Out of Stock" section for the administrator.
Also, the main page, with some categories and the news.
Last, the "add to cart" and "remove from cart" are now with AJAX.